+++ src/Entity/GrantedActions.php
@@ -6,32 +6,29 @@ namespace Dbp\Relay\AuthorizationBundle\Entity;
 
 use ApiPlatform\Metadata\ApiProperty;
 use ApiPlatform\Metadata\ApiResource;
-use ApiPlatform\Metadata\Get;
-use ApiPlatform\Metadata\GetCollection;
-use ApiPlatform\OpenApi\Model\Operation;
-use Dbp\Relay\AuthorizationBundle\Rest\GrantedActionsProvider;
 use Dbp\Relay\CoreBundle\Exception\ApiError;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Serializer\Annotation\Groups;
 
 #[ApiResource(
     shortName: 'AuthorizationGrantedActions',
-    operations: [
-        new Get(
-            uriTemplate: '/authorization/granted-actions/{identifier}',
-            openapi: new Operation(
-                tags: ['Authorization']
-            ),
-            provider: GrantedActionsProvider::class
-        ),
-        new GetCollection(
-            uriTemplate: '/authorization/granted-actions',
-            openapi: new Operation(
-                tags: ['Authorization']
-            ),
-            provider: GrantedActionsProvider::class
-        ),
-    ],
+    operations: [],
+    //    operations: [
+    //        new Get(
+    //            uriTemplate: '/authorization/granted-actions/{identifier}',
+    //            openapi: new Operation(
+    //                tags: ['Authorization']
+    //            ),
+    //            provider: GrantedActionsProvider::class
+    //        ),
+    //        new GetCollection(
+    //            uriTemplate: '/authorization/granted-actions',
+    //            openapi: new Operation(
+    //                tags: ['Authorization']
+    //            ),
+    //            provider: GrantedActionsProvider::class
+    //        ),
+    //    ],
     normalizationContext: [
         'groups' => ['AuthorizationGrantedActions:output'],
     ],